2003-11-10  Jakub Jelinek  <jakub@redhat.com>
	* config/d30v/d30v.h (ASM_OUTPUT_SKIP): Cast SIZE, ROUNDED, NBYTES
	to int.
	* config/vax/vms.h (ASM_OUTPUT_COMMON): Likewise.
	* config/i386/cygwin.h (ASM_OUTPUT_COMMON): Likewise.
	* config/mips/iris6.h (ASM_OUTPUT_ALIGNED_LOCAL): Likewise.
	* config/sparc/vxsim.h (ASM_OUTPUT_SKIP): Likewise.
	* config/i370/i370.h (ASM_OUTPUT_SKIP): Change s to unsigned
	HOST_WIDE_INT.
	(ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL): Use
	HOST_WIDE_INT_PRINT_UNSIGNED for ROUNDED.
	* config/ptx4.h (ASM_OUTPUT_ALIGNED_COMMON): Use
	HOST_WIDE_INT_PRINT_UNSIGNED for SIZE.
	* config/v850/v850.c (v850_output_aligned_bss,
	v850_output_local): Cast size to unsigned HOST_WIDE_INT.

	* Makefile.in (STRICT_WARN, STRICT2_WARN): Remove -Wtraditional.

2003-06-06  John David Anglin  <dave.anglin@nrc-cnrc.gc.ca>

	* pa.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_ALIGNED_COMMON,
	ASM_OUTPUT_ALIGNED_LOCAL): Cast `SIZE' and `ALIGNED' parameters to
	unsigned HOST_WIDE_INT.
	* pa-pro-end.h (ASM_OUTPUT_ALIGNED_COMMOM, ASM_OUTPUT_ALIGNED_LOCAL):
	Likewise.
	* pa64-hpux.h (ASM_OUTPUT_ALIGNED_COMMON, ASM_OUTPUT_ALIGNED_LOCAL):
	Likewise.

2003-06-04  Kaveh R. Ghazi  <ghazi@caip.rutgers.edu>

	* arm/aout.h (ASM_OUTPUT_SKIP): Fix cast for format specifier warning.

2003-06-04  Kaveh R. Ghazi  <ghazi@caip.rutgers.edu>

	* mips/linux.h (ASM_OUTPUT_ALIGNED_BSS): Fix signed/unsigned warnings.
	* pa/pa-pro-end.h (ASM_OUTPUT_ALIGNED_COMMON): Likewise.
	* pa.h (ASM_OUTPUT_ALIGNED_COMMON): Likewise.

2003-06-04  Kaveh R. Ghazi  <ghazi@caip.rutgers.edu>

	* alpha/elf.h (ASM_OUTPUT_ALIGNED_COMMON): Fix format specifier
	warnings.
	* alpha/vms.h (ASM_OUTPUT_ALIGNED_COMMON): Likewise.
	* arm/aof.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL): Likewise.
	* arm/pe.h (ASM_OUTPUT_COMMON): Likewise.
	* avr.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_SKIP): Likewise.
	* h8300.h (ASM_OUTPUT_COMMON): Likewise.
	* mcore.h (ASM_OUTPUT_ALIGNED_COMMON): Likewise.

2003-05-15  Roger Sayle  <roger@eyesopen.com>

	* config/alpha/alpha.h (ASM_OUTPUT_SKIP): Fix typo.

2003-05-14  Matt Kraai  <kraai@alumni.cmu.edu>

	* config/rs6000/sysv4.h (ASM_OUTPUT_ALIGNED_LOCAL): Cast
	g_switch_value to unsigned HOST_WIDE_INT.

2003-05-13  Ulrich Weigand  <uweigand@de.ibm.com>

	* config/s390/s390.h (ASM_OUTPUT_SPECIAL_POOL_ENTRY): Second
	parameter of ASM_OUTPUT_SKIP must have type unsigned HOST_WIDE_INT.

2003-05-12  David Edelsohn  <edelsohn@gnu.org>

	* config/rs6000/xcoff.h (ASM_OUTPUT_SKIP): Accept HOST_WIDE_INT.
	(ASM_OUTPUT_ALIGNED_COMMON): Same.
	(ASM_OUTPUT_LOCAL): Same.

Mon May 12 15:57:54 CEST 2003  Jan Hubicka  <jh@suse.cz>

	* rs6000/darwin.h (ASM_OUTPUT_COMMON): Accept HOST_WIDE_INT
	* mips/elf.h (ASM_OUTPUT_ALIGNED_BSS): Fix warning.
	* mips/mips.h (ASM_OUTPUT_SKIP):  Fix typo in the previous patch.

Mon May 12 11:32:53 CEST 2003  Jan Hubicka  <jh@suse.cz>

	* expr.h (assemble_static_space): Update prototype.
	* output.h (assemble_zeros, output_constant): Likewise.
	* elfos.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_COMMON): Make it 64bit clean
	* alpha.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_COMMON, ASK_OUTPUT_LOCAL): Make
	it 64bit clean.
	* elf.h (ASM_OTUPUT_SKIP): Likewise.
	* unicosmk.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_COMM): Likewise.
	* arm.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL, ASM_OUTPUT_SKIP):
	Expect HOST_WIDE_INT operand.
	* aout.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL, ASM_OUTPUT_SKIP):
	Expect HOST_WIDE_INT operand.
	* unknown-elf.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL, ASM_OUTPUT_SKIP):
	Expect HOST_WIDE_INT operand.
	* avr.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_SKIP): Expect
	HOST_WIDE_INT operand.
	* c4x.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL, ASM_OUTPUT_SKIP,
	ASM_OUTPUT_BSS): Expect HOST_WIDE_INT operand.
	* aout.h (ASM_OTUPUT_SKIP): Likewise.
	* cris.h (ASM_OTUPUT_SKIP, ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL): Likewise.
	* darwin.h (ASM_OTUPUT_SKIP, ASM_OUTPUT_COMMON): Likewise.
	* dsp16xx.h (ASM_OTUPUT_SKIP): Likewise.
	* frv.h (ASM_OTUPUT_SKIP): Likewise.
	* h8300.h (ASM_OTUPUT_SKIP, ASM_OUTPUT_LOCAL): Likewise.
	* 370.h (ASM_OTUPUT_SKIP, ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL):
	Make it 64bit clean.
	* att.h (ASM_OUTPUT_SKIP): Expect HOST_WIDE_INT operand.
	* bsd.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL):
	Make it 64bit clean.
	* darwin.h (ASM_OUTPUT_SKIP): Make it 64bit clean..
	* sco5.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL): Expect
	HOST_WIDE_INT operand
	* svr3gas.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL):
	Expect HOST_WIDE_INT operand
	* sysv3.h (ASM_OUTPUT_SKIP): Expect HOST_WIDE_INT operand
	* i960.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_COMMON,
	ASM_OUTPUT_ALIGNED_LOCAL): Expect HOST_WIDE_INT operand
	* ip2k.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL, ASM_OUTPUT_SKIP):
	Likewise.
	* m32r.h (ASM_OUTPUT_COMMON): Likewise.
	* 3b1.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL, ASM_OUTPUT_SKIP):
	Likewise.
	* amix.h (ASM_OUTPUT_ALIGNED_LOCAL): Likewise.
	* crds.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_ALIGNED_LOCAL): Likewise.
	* hp320.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL,
	ASM_OUTPUT_SKIP): Likewise.
	* m68k.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL, ASM_OUTPUT_SKIP):
	Likewise.
	* m68kelf.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL,
	ASM_OUTPUT_SKIP): Likewise.
	* m68kv4.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL,
	ASM_OUTPUT_SKIP): Likewise.
	* mot3300.h (ASM_OUTPUT_LOCAL, ASM_OUTPUT_SKIP): Likewise.
	* netbsd-elf.h (ASM_OUTPUT_LOCAL, ASM_OUTPUT_COMMON,
	ASM_OUTPUT_SKIP): Likewise.
	* sgs.h (ASM_OUTPUT_SKIP): Likewise.
	* tower-as.h (ASM_OUTPUT_LOCAL, ASM_OUTPUT_COMMON,
	ASM_OUTPUT_SKIP): Likewise.
	* m88k.h (ASM_OUTPUT_LOCAL, ASM_OUTPUT_COMMON, ASM_OUTPUT_SKIP):
	Likewise.
	* mcore.h (ASM_OUTPUT_LOCAL, ASM_OUTPUT_BSS, ASM_OUTPUT_SKIP): Likewise.
	* iris.h (ASM_OUTPUT_LOCAL): Likewise.
	* mips.h (ASM_OUTPUT_LOCAL, ASM_OUTPUT_SKIP): Likewise.
	* ns32k.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL,
	ASM_OUTPUT_SKIP): Make it 64bit clean.
	* pa-pro-end.h (ASM_OUTPUT_LOCAL, ASM_OUTPUT_ALIGNED_LOCAL):
	Make it 64bit clean.
	* pa.h (ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL, ASM_OUTPUT_ALIGNED_LOCAL,
	ASM_OUTPUT_SKIP): Make it 64bit clean.
	* hpux.h (ASM_OUTPUT_LOCAL, ASM_OUTPUT_ALIGNED_LOCAL): Make it
	64bit clean.
	* romp.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_LOCAL, ASM_OUTPUT_COMMON): Expect
	HOST_WIDE_INT argument
	* s390.h (ASM_OUTPUT_SKIP): Expect HOST_WIDE_INT argument.
	* sh.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_LOCAL, ASM_OUTPUT_COMMON): Expect
	HOST_WIDE_INT argument
	* sol2.h (ASM_OUTPUT_SKIP): HOST_WIDE_INT argument
	* sparc.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_COMMON): HOST_WIDE_INT argument
	* svr3.h (ASM_OUTPUT_SKIP): HOST_WIDE_INT argument
	* vax.h (ASM_OUTPUT_SKIP, ASM_OUTPUT_COMMON, ASM_OUTPUT_LOCAL):
	HOST_WIDE_INT argument
	* vaxv.h (ASM_OUTPUT_LOCAL): HOST_WIDE_INT argument
	* xtensa.h (ASM_OUTPUT_LOCAL): HOST_WIDE_INT argument
	* varasm.c (asm_output_bss, asm_output_aligned_bss,
	asm_emit_uninitialized, assemble_zeros, assemble_static_space):
	HOST_WIDE_INT argument


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gcc-3_3-rhl-branch@74227 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/config/arm/elf.h
@@ -158,7 +158,7 @@ Boston, MA 02111-1307, USA.  */
     {								\
       fprintf (STREAM, "\t.comm\t");				\
       assemble_name (STREAM, NAME);				\
-      fprintf (STREAM, ", %d, %d\n", SIZE, ALIGN);		\
+      fprintf (STREAM, ", %d, %d\n", (int)(SIZE), (int)(ALIGN));\
     }								\
   while (0)
