+++ gcc/config/i386/cygwin.h
@@ -287,7 +287,7 @@ do {							\
       fprintf ((STREAM), "\t.comm\t"); 			\
       assemble_name ((STREAM), (NAME));			\
       fprintf ((STREAM), ", %d\t%s %d\n",		\
-	       (ROUNDED), ASM_COMMENT_START, (SIZE));	\
+	       (ROUNDED), ASM_COMMENT_START, (int)(SIZE));	\
     }							\
 } while (0)
